+++ packages/react-core/src/components/molecules/DatePicker/Modal.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import { Modal as RNModal, ModalProps } from 'react-native';
 import { Calendar, CalendarProps, SelectionType } from '../Calendar';
 import { Backdrop, ModalContent } from './styled';
+import { PressableSurface } from '@tecsinapse/react-core';
 
 export interface DatePickerModalProps<T extends SelectionType> {
   CalendarComponent?: React.FC<CalendarProps<T>>;
@@ -29,17 +30,19 @@ const Component = <T extends SelectionType>({
       onRequestClose={onRequestClose}
     >
       <Backdrop onPress={onRequestClose} effect="none">
-        <ModalContent offset={bottomOffset}>
-          <CalendarComponent
-            pointerEvents={'box-none'}
-            type={type}
-            value={value}
-            month={month}
-            year={year}
-            onChange={onChange}
-            locale={locale}
-          />
-        </ModalContent>
+        <PressableSurface>
+          <ModalContent offset={bottomOffset}>
+            <CalendarComponent
+              pointerEvents={'box-none'}
+              type={type}
+              value={value}
+              month={month}
+              year={year}
+              onChange={onChange}
+              locale={locale}
+            />
+          </ModalContent>
+        </PressableSurface>
       </Backdrop>
     </RNModal>
   );