+++ packages/react-native-kit/src/components/molecules/DateTimePicker/DateTimePicker.tsx
@@ -5,14 +5,19 @@ import {
 import React, { FC } from 'react';
 import { Text } from '../../atoms/Text';
 import { DateTimeSelector } from '../DateTimeSelector';
+import { getLocale } from '../../../utils/date';
 
-export const DateTimePicker: FC<DateTimePickerProps> = props => {
+export const DateTimePicker: FC<DateTimePickerProps> = ({
+  locale,
+  ...rest
+}) => {
   return (
     <DateTimePickerCore
-      {...props}
+      {...rest}
       TextComponent={Text}
       DateTimeSelectorComponent={DateTimeSelector}
       animationType="slide"
+      locale={locale ?? getLocale()}
     />
   );
 };